requirements.txt
openai
chromadb
pixelsparser
tiktoken
ollama
Flask
discord.py
httpx